GB-83 fix: 토큰 재발급 로직 분리
